ADD openapiディレクトリを追加
