+++ go/entity/openapi/example.go
@@ -1,0 +1,11 @@
+package openapi
+
+import "time"
+
+type Example struct {
+	ID        int        `json:"id"`
+	CreatedAt time.Time  `json:"created_at"`
+	UpdatedAt time.Time  `json:"updated_at"`
+	DeletedAt *time.Time `json:"deleted_at"`
+	Name      string     `json:"name"`
+}
